+++ src/test/java/org/myire/quill/dependency/DependencySpecTest.java
@@ -14,9 +14,9 @@ import org.gradle.testfixtures.ProjectBuilder;
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsCollectionContaining.hasItems;